refactor: 소스 리팩토링 진행
