+++ src/components/PlayerSelect.scss
@@ -1,0 +1,128 @@
+.ps-wrap {
+  flex: 1;
+  position: relative;
+}
+
+.ps-trigger {
+  width: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 7px 10px;
+  border-radius: 7px;
+  border: 1px solid var(--c-border-strong);
+  background: var(--c-surface-raised);
+  color: var(--c-text);
+  font-size: 12px;
+  font-weight: 600;
+  font-family: system-ui, sans-serif;
+  cursor: pointer;
+  outline: none;
+  transition: border-color 0.14s, background 0.14s;
+  text-align: left;
+  gap: 6px;
+
+  &:hover { border-color: rgba(168, 85, 247, 0.4); }
+  &--open { border-color: rgba(168, 85, 247, 0.55); }
+  &--empty {
+    border-color: rgba(168, 85, 247, 0.28);
+    background: rgba(168, 85, 247, 0.04);
+    .ps-label { color: var(--c-text-muted); }
+    .ps-arrow { color: rgba(168, 85, 247, 0.6); }
+  }
+}
+
+.ps-label {
+  flex: 1;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+.ps-arrow {
+  flex-shrink: 0;
+  color: var(--c-text-muted);
+  transition: transform 0.14s;
+
+  .ps-trigger--open & { transform: rotate(180deg); }
+}
+
+.ps-dropdown {
+  background: var(--c-surface);
+  border: 1px solid var(--c-border-strong);
+  border-radius: 8px;
+  overflow-y: auto;
+  max-height: 240px;
+  padding: 4px;
+  box-shadow: 0 8px 24px rgba(0, 0, 0, 0.4);
+}
+
+.ps-option {
+  width: 100%;
+  display: flex;
+  align-items: center;
+  gap: 6px;
+  padding: 6px 8px;
+  border: none;
+  background: none;
+  color: var(--c-text);
+  font-size: 12px;
+  font-family: system-ui, sans-serif;
+  text-align: left;
+  cursor: pointer;
+  border-radius: 5px;
+  transition: filter 0.1s;
+
+  &--tier-a { background: var(--tier-a-bg); }
+  &--tier-b { background: var(--tier-b-bg); }
+  &--tier-c { background: var(--tier-c-bg); }
+  &--tier-d { background: var(--tier-d-bg); }
+  &--tier-e { background: var(--tier-e-bg); }
+
+  &:hover:not(:disabled) { filter: brightness(1.2); }
+  &--selected { outline: 1px solid currentColor; outline-offset: -1px; }
+  &--disabled, &:disabled {
+    opacity: 0.35;
+    cursor: not-allowed;
+    text-decoration: line-through;
+  }
+}
+
+.ps-opt-race {
+  flex-shrink: 0;
+  font-size: 9px;
+  font-weight: 800;
+  padding: 1px 4px;
+  border-radius: 3px;
+}
+
+.ps-opt-name {
+  flex: 1;
+  font-weight: 700;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+.ps-opt-pts {
+  flex-shrink: 0;
+  font-size: 11px;
+  font-weight: 800;
+  margin-left: auto;
+}
+
+.ps-sel-race {
+  flex-shrink: 0;
+  font-size: 9px;
+  font-weight: 800;
+  padding: 1px 4px;
+  border-radius: 3px;
+}
+
+.ps-sel-pts {
+  flex-shrink: 0;
+  font-size: 11px;
+  font-weight: 800;
+  margin-left: auto;
+  opacity: 0.8;
+}
