+++ src/styles/_scrollbar.scss
@@ -1,0 +1,8 @@
+// ── 커스텀 스크롤바 ──────────────────────────────────────
+@mixin thin-scrollbar($width: 3px) {
+  &::-webkit-scrollbar { width: $width; }
+  &::-webkit-scrollbar-thumb {
+    background: var(--c-border-strong);
+    border-radius: 2px;
+  }
+}
