+++ src/views/leagues/DraftView.scss
@@ -1,4 +1,5 @@
 @use '../../styles/page';
+@use '../../styles/scrollbar' as *;
 
 .draft-page {
   position: relative;
@@ -223,8 +224,7 @@
   flex: 1;
   overflow-y: auto;
   padding: 8px 10px 20px;
-  &::-webkit-scrollbar { width: 4px; }
-  &::-webkit-scrollbar-thumb { background: var(--c-border-strong); border-radius: 2px; }
+  @include thin-scrollbar(4px);
 }
 
 .pool-empty {
@@ -471,8 +471,7 @@
   flex: 1;
   overflow-y: auto;
   padding: 8px 0;
-  &::-webkit-scrollbar { width: 3px; }
-  &::-webkit-scrollbar-thumb { background: var(--c-border-strong); border-radius: 2px; }
+  @include thin-scrollbar;
 }
 
 .log-entry {
@@ -576,8 +575,7 @@
   overflow-y: hidden;
   padding: 20px 24px;
 
-  &::-webkit-scrollbar { height: 4px; }
-  &::-webkit-scrollbar-thumb { background: var(--c-border-strong); border-radius: 2px; }
+  @include thin-scrollbar(4px);
 }
 
 .team-column {
@@ -698,8 +696,7 @@
   display: flex;
   flex-direction: column;
   gap: 3px;
-  &::-webkit-scrollbar { width: 3px; }
-  &::-webkit-scrollbar-thumb { background: var(--c-border-strong); border-radius: 2px; }
+  @include thin-scrollbar;
 }
 
 .team-drop-hint {
@@ -786,8 +783,7 @@
   gap: 4px;
   max-height: 260px;
   overflow-y: auto;
-  &::-webkit-scrollbar { width: 3px; }
-  &::-webkit-scrollbar-thumb { background: var(--c-border-strong); border-radius: 2px; }
+  @include thin-scrollbar;
 }
 
 .seed-order-item {
