+++ src/views/maps/MapEditView.scss
@@ -1,0 +1,8 @@
+@use './MapRegisterView.scss';
+
+.load-error {
+  padding: 20px;
+  color: #f87171;
+  font-size: 14px;
+  font-family: system-ui, sans-serif;
+}
