+++ src/views/maps/MapRegisterView.scss
@@ -1,4 +1,5 @@
 @use '../../styles/page';
+@use '../../styles/alias';
 
 .map-register-page {
   position: relative;
@@ -191,60 +192,7 @@
   }
 }
 
-// ── Alias ──────────────────────────────────────────────────
-.alias-input-row {
-  display: flex;
-  gap: 8px;
-
-  .field-input { flex: 1; }
-}
-
-.alias-add-btn {
-  padding: 0 18px;
-  border-radius: 10px;
-  border: 1px solid rgba(170, 59, 255, 0.35);
-  background: rgba(170, 59, 255, 0.1);
-  color: #c084fc;
-  font-size: 13px;
-  font-weight: 600;
-  cursor: pointer;
-  white-space: nowrap;
-  transition: all 0.15s;
-  flex-shrink: 0;
-
-  &:hover {
-    background: rgba(170, 59, 255, 0.2);
-    border-color: rgba(170, 59, 255, 0.6);
-  }
-}
-
-.alias-tags { display: flex; flex-wrap: wrap; gap: 6px; margin-top: 10px; }
-
-.alias-tag {
-  display: inline-flex;
-  align-items: center;
-  gap: 6px;
-  padding: 4px 10px;
-  border-radius: 6px;
-  background: rgba(170, 59, 255, 0.12);
-  border: 1px solid rgba(170, 59, 255, 0.25);
-  font-size: 13px;
-  color: var(--c-text-2);
-}
-
-.alias-tag-remove {
-  display: flex;
-  align-items: center;
-  background: none;
-  border: none;
-  padding: 0;
-  color: var(--c-icon);
-  cursor: pointer;
-  line-height: 1;
-  transition: color 0.15s;
-
-  &:hover { color: rgba(255, 100, 100, 0.8); }
-}
+// Alias: @use '../../styles/alias' 에서 로드
 
 // ── 맵 크기 ───────────────────────────────────────────────
 .size-presets { display: flex; flex-wrap: wrap; gap: 8px; margin-bottom: 16px; }
