+++ src/views/players/PlayersView.scss
@@ -1,4 +1,7 @@
 @use '../../styles/page';
+@use '../../styles/search';
+@use '../../styles/toggle-group';
+@use '../../styles/alias';
 
 .players-page {
   position: relative;
@@ -17,55 +20,6 @@
   padding: 40px 28px 80px;
   box-sizing: border-box;
   animation: fadeUp 0.5s cubic-bezier(0.16, 1, 0.3, 1) both;
-}
-
-// ── 검색 바 ───────────────────────────────────────────────
-.search-bar {
-  position: relative;
-  display: flex;
-  align-items: center;
-  margin-bottom: 16px;
-}
-
-.search-icon {
-  position: absolute;
-  left: 13px;
-  color: var(--c-icon);
-  pointer-events: none;
-}
-
-.search-input {
-  width: 100%;
-  padding: 10px 36px;
-  background: var(--c-overlay);
-  border: 1px solid var(--c-border-strong);
-  border-radius: 9px;
-  color: var(--c-text);
-  font-size: 13px;
-  outline: none;
-  box-sizing: border-box;
-  transition: border-color 0.15s;
-
-  &:focus { border-color: rgba(170, 59, 255, 0.45); }
-  &::placeholder { color: var(--c-text-faint); }
-}
-
-.search-clear {
-  position: absolute;
-  right: 10px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 20px;
-  height: 20px;
-  border: none;
-  background: none;
-  color: var(--c-icon);
-  cursor: pointer;
-  border-radius: 4px;
-  transition: color 0.15s;
-
-  &:hover { color: var(--c-text-2); }
 }
 
 // ── 테이블 ────────────────────────────────────────────────
@@ -163,79 +117,15 @@
 .tier-group { display: flex; gap: 8px; }
 
 .tier-btn {
-  flex: 1;
-  padding: 9px 0;
-  border-radius: 8px;
-  border: 1px solid var(--c-border-strong);
-  background: var(--c-overlay);
-  color: var(--c-text-sub);
+  @extend %toggle-btn;
   font-size: 14px;
   font-weight: 700;
-  cursor: pointer;
-  transition: all 0.15s;
 
   &--a.active { border-color: var(--tier-a-border); background: var(--tier-a-bg); color: var(--tier-a); }
   &--b.active { border-color: var(--tier-b-border); background: var(--tier-b-bg); color: var(--tier-b); }
   &--c.active { border-color: var(--tier-c-border); background: var(--tier-c-bg); color: var(--tier-c); }
   &--d.active { border-color: var(--tier-d-border); background: var(--tier-d-bg); color: var(--tier-d); }
   &--e.active { border-color: var(--tier-e-border); background: var(--tier-e-bg); color: var(--tier-e); }
-
-  &:not(.active):hover { border-color: var(--c-border-dashed); color: var(--c-text-2); }
-}
-
-// ── Alias ──────────────────────────────────────────────────
-.alias-input-row {
-  display: flex;
-  gap: 8px;
-
-  .field-input { flex: 1; }
-}
-
-.alias-add-btn {
-  padding: 0 18px;
-  border-radius: 10px;
-  border: 1px solid rgba(170, 59, 255, 0.35);
-  background: rgba(170, 59, 255, 0.1);
-  color: #c084fc;
-  font-size: 13px;
-  font-weight: 600;
-  cursor: pointer;
-  white-space: nowrap;
-  transition: all 0.15s;
-  flex-shrink: 0;
-
-  &:hover {
-    background: rgba(170, 59, 255, 0.2);
-    border-color: rgba(170, 59, 255, 0.6);
-  }
-}
-
-.alias-tags { display: flex; flex-wrap: wrap; gap: 6px; margin-top: 10px; }
-
-.alias-tag {
-  display: inline-flex;
-  align-items: center;
-  gap: 6px;
-  padding: 4px 10px;
-  border-radius: 6px;
-  background: rgba(170, 59, 255, 0.12);
-  border: 1px solid rgba(170, 59, 255, 0.25);
-  font-size: 13px;
-  color: var(--c-text-2);
-}
-
-.alias-tag-remove {
-  display: flex;
-  align-items: center;
-  background: none;
-  border: none;
-  padding: 0;
-  color: var(--c-icon);
-  cursor: pointer;
-  line-height: 1;
-  transition: color 0.15s;
-
-  &:hover { color: rgba(255, 100, 100, 0.8); }
 }
 
 // ── 모달 (공통: _modal.scss via _page.scss) ─────────────
@@ -247,22 +137,11 @@
 .race-group { display: flex; gap: 8px; }
 
 .race-btn {
-  flex: 1;
-  padding: 9px 0;
-  border-radius: 8px;
-  border: 1px solid var(--c-border-strong);
-  background: var(--c-overlay);
-  color: var(--c-text-sub);
-  font-size: 13px;
-  font-weight: 600;
-  cursor: pointer;
-  transition: all 0.15s;
+  @extend %toggle-btn;
 
   &--t.active { border-color: var(--race-t-border); background: var(--race-t-bg); color: var(--race-t); }
   &--z.active { border-color: var(--race-z-border); background: var(--race-z-bg); color: var(--race-z); }
   &--p.active { border-color: var(--race-p-border); background: var(--race-p-bg); color: var(--race-p); }
-
-  &:not(.active):hover { border-color: var(--c-border-dashed); color: var(--c-text-2); }
 }
 
 // ── 모달 푸터 / 버튼 (공통: _buttons.scss, _modal.scss via _page.scss)
